Moving ErrorDisplayableDecorator to Common folder and its extensions to their corresponding modules
